Add Camera Rotation
Added public dot product, cross product, and rotation calculation methods to Data::Position.
Added public right vector calculation method to Display::FreeGlut::Camera.
Added public movement and rotation methods to Display::FreeGlut::Camera.
Changed main use new methods to control camera rotation and position.
Changed main to exit on escape press and reset camera on space press.
Fixed type conversion in Data::Model.

+++ model.cpp
@@ -92,7 +92,7 @@ const Data::Position<float> Data::Model::calcCenter() const
 		center += *it;
 	}
 
-	return center / this->vertices.size();
+	return center / (float)this->vertices.size();
 }
 
 const float Data::Model::calcXRadius() const
